requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0

# NLP — Python 3.13 compatible versions
nltk==3.8.1
spacy==3.8.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# ML — numpy 2.x required for Python 3.13
scikit-learn==1.4.2
numpy==2.0.2
pandas==2.2.2

# Document parsing
pymupdf==1.24.14
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pytest==7.4.3
